New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs:   Put in Release Notes: No              Module(s): images            Description: Fixed type ambiguity problem which prevented 			this from compiling on 32 bit Linux.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@15209 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casacore/images/IO/AsciiRegionFileParser.cc
@@ -407,7 +407,7 @@ AsciiRegionFileParser::_getCurrentParamSet(
 		String keyword = consumeMe.substr(0, equalPos);
 		keyword.trim();
 		keyword.downcase();
-		consumeMe.del(0, equalPos + 1);
+		consumeMe.del(0, (Int)equalPos + 1);
 		consumeMe.trim();
 		if (keyword == "label") {
 			key = LABEL;
@@ -902,7 +902,7 @@ Vector<Quantity> AsciiRegionFileParser::_extractTwoQuantityPairs(
 		}
 	}
 
-    consumeMe.del(0, end + 1);
+	consumeMe.del(0, (Int)end + 1);
 	return quantities;
 }
